КТФ: Добавлены окна текущих параметров и добавлены 2 окна конфигурирования

+++ config/config84.h
@@ -80,7 +80,7 @@ public:
         quint32		baud;
         quint8		parity;
         quint8		stopbit;
-        quint16     adrMB;
+        quint8     adrMB;
         quint8      isNTP;
     } Com;
